+++ src/index.css
@@ -1,4 +1,14 @@
 /* write common styles only here */
+:root {
+  --primary-color: #223f64;
+  --secondray-color: #3f88d8;
+  --text-primary-color: #24384f;
+  --font-family: "Poppins", sans-serif;
+  --text-secondray-color: white;
+  --h1-font-size: 32px;
+  --p-font-size: 24px;
+  --error: red;
+}
 *{
   padding: 0;
   margin: 0;
@@ -11,7 +21,7 @@ body {
 
 /* TableComponent.css */
 .table-wrapper {
-  overflow-x: auto; /* Enable horizontal scrolling if needed */
+  overflow-x: auto; 
 }
 
 .antd-table {
@@ -21,6 +31,9 @@ body {
   margin-bottom: 20px;
   box-shadow: 0 2px 8px rgba(0, 0, 0, 0.09);
   table-layout: fixed;
+}
+.antd-table a{
+  color: var(--primary-color);
 }
 
 .antd-table th,
@@ -90,13 +103,13 @@ body {
   width: 32px;
   height: 32px;
   border-radius: 4px;
-  color: #333;
+  color: var(--primary-color);
 }
 
 .ant-pagination-item-active a {
-  background-color: #223f64;
+  background-color: var(--primary-color);
   color: #fff;
-  border-color: #223f64;
+  border-color: var(--primary-color);
 }
 
 .ant-pagination-prev,
@@ -110,9 +123,6 @@ body {
   display: none;
 }
 
-.ant-pagination-options {
-  display: none;
-}
 .ant-pagination .ant-pagination-item-active{
   border: none;
 }
@@ -120,8 +130,8 @@ body {
   color: white;
 }
 
-/* invite nad view more button */
-.create-login-btn{
+/* invite nad view more button button*/ 
+.btn {
   padding: 8px 20px;
   border-radius: 6px;
   outline: none;
@@ -129,16 +139,12 @@ body {
   color: black;
   background-color: white;
   border: 1px solid gray;
-}
-.table-view-more{
-  padding: 8px 20px;
-  border-radius: 6px;
+  cursor: pointer;
   outline: none;
   border: none;
-  color: black;
-  background-color: white;
-  border: 1px solid gray;
 }
+
+
 
 /* 
 common layout */
@@ -232,7 +238,7 @@ common layout */
 }
 
 .active{
-  background-color: #223F64;
+  background-color: var(--primary-color);
   border-radius: 0px 15px 15px 0px;
 }
 .setting{
@@ -257,7 +263,7 @@ common layout */
 }
 
 .filter__icon{
-  color: #223F64;
+  color: var(--primary-color);
 }
 
 .upload__btn{
@@ -266,7 +272,7 @@ common layout */
   outline: none;
   border: none;
   color: white;
-  background-color: #223F64;
+  background-color: var(--primary-color);
 }
 .hide{
   visibility: hidden;
