+++ src/pages/taskModule/TaskModule.css
@@ -1,5 +1,24 @@
+:root {
+  --primary-color: #223f64;
+  --secondray-color: #3f88d8;
+  --text-primary-color: #24384f;
+  --font-family: "Poppins", sans-serif;
+  --text-secondray-color: white;
+  --h1-font-size: 32px;
+  --p-font-size: 24px;
+  --error: red;
+}
+.btn.edit-btn{
+  color: white;
+  background-color: var(--primary-color);
+}
+.btn.delete-btn{
+  color: white;
+  background-color: var(--error);
+}
 span.ant-dropdown-trigger {
-  background-color: #223f64;
+
+  background-color: var(--primary-color);
   color: white;
   padding: 5px 10px;
   border-radius: 5px;
