--- src/utils/formatDate.js
@@ -1,7 +1,0 @@
-//sample utilities used inside comps
-
-const formatDate = () => {
-  return "formatdate";
-};
-
-export { formatDate };
